refactor(handle 5xx errors) refactor response handler and rename files PE-1747
HttpRetry now get retryIf and onRetry as injection on its function.
