+++ lib/services/arweave/error/gateway_error.dart
