+++ lib/services/arweave/error/gateway_response_handler.dart
@@ -1,18 +1,13 @@
-import 'package:ardrive/services/arweave/error/error.dart';
+import 'package:ardrive/services/arweave/error/gateway_error.dart';
+import 'package:ardrive/utils/response_handler.dart';
 import 'package:http/http.dart';
-
-/// `handle` verifies if the response has any erros, and if so, should throw
-/// an exception.
-abstract class ResponseHandler {
-  void handle(Response response);
-}
 
 /// Throws a `GatewayError` exception for `statusCode` different from 200
 ///
 class GatewayResponseHandler implements ResponseHandler {
   @override
   void handle(Response response) {
-    if (response.statusCode != 200) {
+    if (response.statusCode >= 200 && response.statusCode < 300) {
       throw GatewayError.fromResponse(response);
     }
   }
